+++ packages/app/src/pages/Rewards/Overview/RecentPayouts/ActiveGraph.tsx
