+++ packages/app/src/pages/Rewards/Overview/RecentPayouts/InactiveGraph.tsx
